+++ stanford_media.module
@@ -289,5 +289,5 @@ function stanford_media_save_file_managed_media(array $element, FormStateInterfa
  */
 function stanford_media_ckeditor_css_alter(array &$css, Editor $editor) {
   // Add special styles to wysiwyg embeded videos.
-  $css[] = drupal_get_path('module','stanford_media') . '/css/stanford_media.video_preview.css';
+  $css[] = drupal_get_path('module', 'stanford_media') . '/css/stanford_media.video_preview.css';
 }
